+++ cb-trading-db.py
@@ -5,67 +5,97 @@ import secrets
 import json
 from cryptography.hazmat.primitives import serialization
 from collections import deque
+import psycopg2
+from psycopg2.extras import Json
+from decimal import Decimal
 
-# Load API credentials & trading settings from config.json
+# Load configuration from config.json
 with open("config.json", "r") as f:
     config = json.load(f)
 
 key_name = config["name"]
 key_secret = config["privateKey"]
-crypto_symbols = config.get("crypto_symbols", ["ETH", "XRP", "DOGE", "SOL"])  # List of cryptocurrencies to trade
 quote_currency = "USDC"
-buy_threshold = config.get("buy_percentage", -3)  # % drop to buy
-sell_threshold = config.get("sell_percentage", 3)  # % rise to sell
 trade_percentage = config.get("trade_percentage", 10)  # % of available balance to trade
 stop_loss_percentage = config.get("stop_loss_percentage", -10)  # Stop-loss threshold
-volatility_window = config.get("volatility_window", 10)  # Window for calculating volatility
-trend_window = config.get("trend_window", 20)  # Window for calculating moving average
-
-# Minimum order sizes (adjust based on Coinbase requirements for each cryptocurrency)
-min_order_sizes = {
-    "ETH": {"buy": 0.01, "sell": 0.0001},  # ETH minimums
-    "XRP": {"buy": 0.01, "sell": 1},       # XRP minimums (example values, check Coinbase)
-    "DOGE": {"buy": 0.01, "sell": 1},      # DOGE minimums (example values, check Coinbase)
-    "SOL": {"buy": 0.01, "sell": 0.01},    # SOL minimums (example values, check Coinbase)
-}
 
 request_host = "api.coinbase.com"
 
+# Load coin-specific settings
+coins_config = config.get("coins", {})
+crypto_symbols = [symbol for symbol, settings in coins_config.items() if settings.get("enabled", False)]
+
 # Initialize price_history with maxlen equal to the larger of volatility_window and trend_window
-price_history_maxlen = max(volatility_window, trend_window)
+price_history_maxlen = max(
+    max(settings.get("volatility_window", 10) for settings in coins_config.values()),
+    max(settings.get("trend_window", 20) for settings in coins_config.values())
+)
 
-# Load or initialize state
-state_file = "state.json"
-try:
-    with open(state_file, "r") as f:
-        crypto_data = json.load(f)
-        # Convert price_history back to deque
-        for symbol in crypto_symbols:
-            if symbol in crypto_data:
-                crypto_data[symbol]["price_history"] = deque(crypto_data[symbol]["price_history"], maxlen=price_history_maxlen)
-except FileNotFoundError:
-    crypto_data = {
-        symbol: {
-            "price_history": deque(maxlen=price_history_maxlen),
-            "initial_price": None,
-            "total_trades": 0,
-            "total_profit": 0.0,
-        }
-        for symbol in crypto_symbols
-    }
+# Database connection parameters
+DB_HOST = config["database"]["host"]
+DB_PORT = config["database"]["port"]
+DB_NAME = config["database"]["name"]
+DB_USER = config["database"]["user"]
+DB_PASSWORD = config["database"]["password"]
 
-def save_state():
-    """Save the current state to a file."""
-    # Convert deque to list for JSON serialization
-    state_to_save = {
-        symbol: {
-            **data,
-            "price_history": list(data["price_history"]),
-        }
-        for symbol, data in crypto_data.items()
-    }
-    with open(state_file, "w") as f:
-        json.dump(state_to_save, f, indent=2)
+def get_db_connection():
+    """Connect to the PostgreSQL database."""
+    conn = psycopg2.connect(
+        host=DB_HOST,
+        port=DB_PORT,
+        database=DB_NAME,
+        user=DB_USER,
+        password=DB_PASSWORD
+    )
+    return conn
+
+def save_state(symbol, price_history, initial_price, total_trades, total_profit):
+    """Save the trading state to the PostgreSQL database."""
+    conn = get_db_connection()
+    cursor = conn.cursor()
+    try:
+        cursor.execute("""
+        INSERT INTO trading_state (symbol, price_history, initial_price, total_trades, total_profit)
+        VALUES (%s, %s, %s, %s, %s)
+        ON CONFLICT (symbol) DO UPDATE
+        SET price_history = EXCLUDED.price_history,
+            initial_price = EXCLUDED.initial_price,
+            total_trades = EXCLUDED.total_trades,
+            total_profit = EXCLUDED.total_profit
+        """, (symbol, Json(price_history), initial_price, total_trades, total_profit))
+        conn.commit()
+    except Exception as e:
+        print(f"Error saving state to database: {e}")
+    finally:
+        cursor.close()
+        conn.close()
+
+def load_state(symbol):
+    """Load the trading state from the PostgreSQL database."""
+    conn = get_db_connection()
+    cursor = conn.cursor()
+    try:
+        cursor.execute("SELECT price_history, initial_price, total_trades, total_profit FROM trading_state WHERE symbol = %s", (symbol,))
+        row = cursor.fetchone()
+        if row:
+            # Convert decimal.Decimal to float
+            price_history = row[0]
+            initial_price = float(row[1]) if isinstance(row[1], Decimal) else row[1]
+            total_trades = int(row[2])
+            total_profit = float(row[3]) if isinstance(row[3], Decimal) else row[3]
+            return {
+                "price_history": price_history,
+                "initial_price": initial_price,
+                "total_trades": total_trades,
+                "total_profit": total_profit,
+            }
+        return None
+    except Exception as e:
+        print(f"Error loading state from database: {e}")
+        return None
+    finally:
+        cursor.close()
+        conn.close()
 
 def build_jwt(uri):
     """Generate a JWT token for Coinbase API authentication."""
@@ -147,16 +177,18 @@ def place_order(crypto_symbol, side, amount):
         }
     }
     
+    min_order_sizes = coins_config[crypto_symbol]["min_order_sizes"]
+    
     if side == "BUY":
         rounded_amount = round(amount, 2)  # USDC should have 2 decimal places
-        if rounded_amount < min_order_sizes[crypto_symbol]["buy"]:
-            print(f"🚫 Buy order too small: ${rounded_amount:.2f} (minimum: ${min_order_sizes[crypto_symbol]['buy']:.2f})")
+        if rounded_amount < min_order_sizes["buy"]:
+            print(f"🚫 Buy order too small: ${rounded_amount:.2f} (minimum: ${min_order_sizes['buy']:.2f})")
             return False
         order_data["order_configuration"]["market_market_ioc"]["quote_size"] = str(rounded_amount)
     else:  # SELL
         rounded_amount = round(amount, 6)  # Cryptocurrency amount precision
-        if rounded_amount < min_order_sizes[crypto_symbol]["sell"]:
-            print(f"🚫 Sell order too small: {rounded_amount:.6f} {crypto_symbol} (minimum: {min_order_sizes[crypto_symbol]['sell']:.6f} {crypto_symbol})")
+        if rounded_amount < min_order_sizes["sell"]:
+            print(f"🚫 Sell order too small: {rounded_amount:.6f} {crypto_symbol} (minimum: {min_order_sizes['sell']:.6f} {crypto_symbol})")
             return False
         order_data["order_configuration"]["market_market_ioc"]["base_size"] = str(rounded_amount)
 
@@ -182,11 +214,14 @@ def calculate_volatility(price_history):
     price_changes = [(price_history[i] - price_history[i - 1]) / price_history[i - 1] for i in range(1, len(price_history))]
     return sum(price_changes) / len(price_changes)  # Average price change
 
-def calculate_moving_average(price_history):
+def calculate_moving_average(price_history, trend_window):
     """Calculate the moving average of prices."""
     if len(price_history) < trend_window:
         return None
     return sum(price_history) / len(price_history)
+
+# Initialize crypto_data as a global variable
+crypto_data = {}
 
 def trading_bot():
     """Monitors multiple cryptocurrencies and trades based on percentage changes."""
@@ -194,12 +229,21 @@ def trading_bot():
 
     # Initialize initial prices for all cryptocurrencies
     for symbol in crypto_symbols:
-        if crypto_data[symbol]["initial_price"] is None:
+        state = load_state(symbol)
+        if state:
+            crypto_data[symbol] = state
+        else:
             initial_price = get_crypto_price(symbol)
             if not initial_price:
                 print(f"🚨 Failed to fetch initial {symbol} price. Skipping {symbol}.")
                 continue
-            crypto_data[symbol]["initial_price"] = initial_price
+            crypto_data[symbol] = {
+                "price_history": deque(maxlen=price_history_maxlen),
+                "initial_price": initial_price,
+                "total_trades": 0,
+                "total_profit": 0.0,
+            }
+            save_state(symbol, list(crypto_data[symbol]["price_history"]), initial_price, 0, 0.0)
             print(f"🔍 Monitoring {symbol}... Initial Price: ${initial_price:.2f}")
 
     while True:
@@ -216,71 +260,53 @@ def trading_bot():
             price_change = ((current_price - crypto_data[symbol]["initial_price"]) / crypto_data[symbol]["initial_price"]) * 100
             print(f"📈 {symbol} Price: ${current_price:.2f} ({price_change:.2f}%)")
 
-            # Log price history for debugging
-            print(f"🔧 Debug for {symbol}:")
-            print(f"  - Price History: {list(crypto_data[symbol]['price_history'])}")
-            print(f"  - Price History Length: {len(crypto_data[symbol]['price_history'])}")
+            # Get coin-specific settings
+            coin_settings = coins_config[symbol]
+            buy_threshold = coin_settings["buy_percentage"]
+            sell_threshold = coin_settings["sell_percentage"]
+            volatility_window = coin_settings["volatility_window"]
+            trend_window = coin_settings["trend_window"]
 
             # Calculate volatility and moving average
             volatility = calculate_volatility(crypto_data[symbol]["price_history"])
-            moving_avg = calculate_moving_average(crypto_data[symbol]["price_history"])
-
-            # Log moving average calculation
-            if moving_avg is not None:
-                print(f"  - Moving Average: ${moving_avg:.2f}")
-            else:
-                print(f"  - Moving Average: Not enough data (requires at least {trend_window} prices)")
+            moving_avg = calculate_moving_average(crypto_data[symbol]["price_history"], trend_window)
 
             # Adjust thresholds based on volatility
             dynamic_buy_threshold = buy_threshold * (1 + abs(volatility))
             dynamic_sell_threshold = sell_threshold * (1 + abs(volatility))
 
-            # Log dynamic thresholds
-            print(f"  - Dynamic Buy Threshold: {dynamic_buy_threshold:.2f}%")
-            print(f"  - Dynamic Sell Threshold: {dynamic_sell_threshold:.2f}%")
+            # Calculate expected buy/sell prices
+            expected_buy_price = crypto_data[symbol]["initial_price"] * (1 + dynamic_buy_threshold / 100)
+            expected_sell_price = crypto_data[symbol]["initial_price"] * (1 + dynamic_sell_threshold / 100)
+
+            # Log expected prices
+            print(f"📊 Expected Buy Price for {symbol}: ${expected_buy_price:.2f}")
+            print(f"📊 Expected Sell Price for {symbol}: ${expected_sell_price:.2f}")
 
             # Check if the price is close to the moving average
-            if moving_avg:
-                price_deviation = abs(current_price - moving_avg) / moving_avg * 100
-                print(f"  - Price Deviation from Moving Average: {price_deviation:.2f}%")
-                if price_deviation > 2:  # Only trade if price is within 2% of the moving average
-                    print("  ⚠️ Price is too far from the moving average. Skipping trade.")
-                    continue
+            if moving_avg and abs(current_price - moving_avg) < (0.02 * moving_avg):  # Only trade if price is within 2% of the moving average
+                if price_change <= dynamic_buy_threshold and balances[quote_currency] > 0:
+                    buy_amount = (trade_percentage / 100) * balances[quote_currency] / current_price
+                    if buy_amount > 0:
+                        print(f"💰 Buying {buy_amount:.4f} {symbol}!")
+                        if place_order(symbol, "BUY", buy_amount):
+                            crypto_data[symbol]["total_trades"] += 1
+                            crypto_data[symbol]["initial_price"] = current_price  # Reset reference price
 
-            # Calculate expected buy price
-            expected_buy_price = crypto_data[symbol]["initial_price"] * (1 + dynamic_buy_threshold / 100)
-            print(f"📊 Expected Buy Price for {symbol}: ${expected_buy_price:.2f}")
-
-            # Check for buy condition
-            if price_change <= dynamic_buy_threshold and balances[quote_currency] > 0:
-                buy_amount = (trade_percentage / 100) * balances[quote_currency] / current_price
-                print(f"  - Calculated Buy Amount: {buy_amount:.4f} {symbol}")
-                if buy_amount >= min_order_sizes[symbol]["buy"]:
-                    print(f"💰 Buying {buy_amount:.4f} {symbol}!")
-                    if place_order(symbol, "BUY", buy_amount):
-                        crypto_data[symbol]["total_trades"] += 1
-                        crypto_data[symbol]["initial_price"] = current_price  # Reset reference price
-                else:
-                    print(f"🚫 Buy order too small: {buy_amount:.4f} {symbol} (minimum: {min_order_sizes[symbol]['buy']:.4f} {symbol})")
-
-            # Check for sell condition
-            elif price_change >= dynamic_sell_threshold and balances[symbol] > 0:
-                sell_amount = (trade_percentage / 100) * balances[symbol]
-                print(f"  - Calculated Sell Amount: {sell_amount:.4f} {symbol}")
-                if sell_amount >= min_order_sizes[symbol]["sell"]:
-                    print(f"💵 Selling {sell_amount:.4f} {symbol}!")
-                    if place_order(symbol, "SELL", sell_amount):
-                        crypto_data[symbol]["total_trades"] += 1
-                        crypto_data[symbol]["total_profit"] += (current_price - crypto_data[symbol]["initial_price"]) * sell_amount
-                        crypto_data[symbol]["initial_price"] = current_price  # Reset reference price
-                else:
-                    print(f"🚫 Sell order too small: {sell_amount:.4f} {symbol} (minimum: {min_order_sizes[symbol]['sell']:.4f} {symbol})")
+                elif price_change >= dynamic_sell_threshold and balances[symbol] > 0:
+                    sell_amount = (trade_percentage / 100) * balances[symbol]
+                    if sell_amount > 0:
+                        print(f"💵 Selling {sell_amount:.4f} {symbol}!")
+                        if place_order(symbol, "SELL", sell_amount):
+                            crypto_data[symbol]["total_trades"] += 1
+                            crypto_data[symbol]["total_profit"] += (current_price - crypto_data[symbol]["initial_price"]) * sell_amount
+                            crypto_data[symbol]["initial_price"] = current_price  # Reset reference price
 
             # Log performance for each cryptocurrency
             print(f"📊 {symbol} Performance - Total Trades: {crypto_data[symbol]['total_trades']} | Total Profit: ${crypto_data[symbol]['total_profit']:.2f}")
 
-        # Save state after each iteration
-        save_state()
+            # Save state after each coin's update
+            save_state(symbol, list(crypto_data[symbol]["price_history"]), crypto_data[symbol]["initial_price"], crypto_data[symbol]["total_trades"], crypto_data[symbol]["total_profit"])
 
 if __name__ == "__main__":
     trading_bot()
